pyproject.toml
[tool.poetry]
name = "user-crud-assignment"
version = "0.1.0"
description = ""
authors = ["Shubham Singhal <singhal121997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.9.2"}


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
aiosqlite = "^0.20.0"
pytest-cov = "^6.0.0"
coverage = "^7.6.7"
gevent = "^24.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
concurrency = ["gevent"]
